+++ nichtparasoup/config/__init__.py
@@ -14,14 +14,14 @@ _defaults = None  # type: Optional[Dict[str, Any]]
 
 def get_imagecrawler(config_crawler: Dict[str, Any]) -> BaseImageCrawler:
     from nichtparasoup.imagecrawler import get_imagecrawlers
-    imagecrawler_class = get_imagecrawlers().get_class(config_crawler['type'])
+    imagecrawler_class = get_imagecrawlers().get_class(config_crawler['name'])
     if not imagecrawler_class:
-        raise ValueError('unknown crawler type {!r}'.format(config_crawler['type']))
+        raise ValueError('unknown crawler name {!r}'.format(config_crawler['name']))
     try:
         imagecrawler_obj = imagecrawler_class(**config_crawler['config'])
     except Exception as e:
-        raise Exception('failed setup of {type!r} with config {config!r}'.format(
-            type=config_crawler['type'], config=config_crawler['config'])) from e
+        raise Exception('failed setup crawler {name!r} of type {type!r} with config {config!r}'.format(
+            name=config_crawler['name'], type=imagecrawler_class, config=config_crawler['config'])) from e
     return imagecrawler_obj