+++ tests/test_nichtparasoup/test_cmdline/test_main.py
@@ -1,0 +1,7 @@
+import unittest
+
+
+class MainTest(unittest.TestCase):
+
+    def test_main(self) -> None:
+        self.skipTest('TODO')