+++ tests/test_nichtparasoup/test_core/mockable_imagecrawler.py
@@ -14,7 +14,6 @@ class MockableImageCrawler(BaseImageCrawler):
         return ImageCrawlerInfo(
             description='a mock',
             config=dict(),
-            version='mock',
         )
 
     @classmethod